+++ src/carousel-gallery-block/edit.js
@@ -780,6 +780,16 @@ export default function Edit( { clientId, attributes, setAttributes } ) {
 							{ attributes, setAttributes }
 						) }
 					</PanelBody>
+					<PanelBody
+						title={ __( 'Watermark Overlay', 'folioblocks' ) }
+						initialOpen={ false }
+					>
+						{ applyFilters(
+							'folioBlocks.carouselGallery.watermarkControls',
+							imageProFeatureNotice( 'watermarkOverlay' ),
+							{ attributes, setAttributes }
+						) }
+					</PanelBody>
 				</InspectorControls>
 			<InspectorControls group="advanced">
 				{ applyFilters(