+++ src/filmstrip-gallery-block/edit.js
@@ -1285,6 +1285,16 @@ export default function Edit( { attributes, setAttributes, clientId } ) {
 						{ attributes, setAttributes }
 					) }
 				</PanelBody>
+				<PanelBody
+					title={ __( 'Watermark Overlay', 'folioblocks' ) }
+					initialOpen={ false }
+				>
+					{ applyFilters(
+						'folioBlocks.filmstripGallery.watermarkControls',
+						imageProFeatureNotice( 'watermarkOverlay' ),
+						{ attributes, setAttributes }
+					) }
+				</PanelBody>
 			</InspectorControls>
 			<InspectorControls group="advanced">
 				{ applyFilters(